+++ device/thunder_ripple_sdk/src/bootstrap/boot_thunder.rs
@@ -52,7 +52,7 @@ pub async fn boot_thunder(
 ) -> Option<ThunderBootstrapStateWithClient> {
     info!("Booting thunder initiated");
     //by default enabling the thunderBroker
-    let state = if ext_client.get_bool_config("use_with_thunder_broker") {
+    let state = if ext_client.get_bool_config("use_with_thunder_async_client") {
         info!("Using thunder broker");
         let mut extn_client = ext_client.clone();
         let mut gateway_url = url::Url::parse(GATEWAY_DEFAULT).unwrap();
